fixed rental and revenue conflict
