--- src/main/java/data/PurchaseTransactions.json
@@ -1,16 +1,0 @@
-{
-  "1": {
-    "transactionId": 1,
-    "bookId": 101,
-    "bookName": "Effective Java",
-    "soldPrice": 45.99,
-    "date": "2024-07-16"
-  },
-  "2": {
-    "transactionId": 2,
-    "bookId": 102,
-    "bookName": "Clean Code",
-    "soldPrice": 39.99,
-    "date": "2024-07-10"
-  }
-}
