+++ src/main/java/data_access/database_transaction_entry/DataTransactionEntryDataAccessObject.java
@@ -10,23 +10,17 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Date;
-import java.util.Iterator;
 import java.util.Set;
 import java.text.SimpleDateFormat;
-import java.util.Date;
-import data.FilePathConstants;
+
+import data.misc_info.FilePathConstants;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 
-import java.io.FileWriter;
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.Date;
-import java.util.Map;
 
 public class DataTransactionEntryDataAccessObject implements DatabaseTransactionEntryDataAccessInterface{
