+++ src/main/java/interface_adapter/returnorborrow/ReturnOrBorrowPresenter.java
@@ -3,8 +3,10 @@ package interface_adapter.returnorborrow;
 import interface_adapter.RentInformation.borrowbook.BorrowBookViewModel;
 import interface_adapter.RentInformation.returnbook.ReturnBookViewModel;
 import interface_adapter.RentMenu.RentMenuViewModel;
+import interface_adapter.main_menu.MainMenuViewModel;
 import interface_adapter.view.ViewManagerModel;
 import use_case.rent_book.ReturnOrBorrow.RobOutputBoundary;
+import view.views.MainMenuView;
 
 public class ReturnOrBorrowPresenter implements RobOutputBoundary {
     private final ReturnOrBorrowViewModel returnOrBorrowViewModel;
@@ -12,18 +14,19 @@ public class ReturnOrBorrowPresenter implements RobOutputBoundary {
     private final RentMenuViewModel rentMenuViewModel;
     private final ReturnBookViewModel returnBookViewModel;
     private final BorrowBookViewModel borrowBookViewModel;
+    private final MainMenuViewModel mainMenuViewModel;
 
-    public ReturnOrBorrowPresenter(ReturnOrBorrowViewModel returnOrBorrowViewModel, ViewManagerModel viewManagerModel, RentMenuViewModel rentMenuViewModel, ReturnBookViewModel returnBookViewModel, BorrowBookViewModel borrowBookViewModel) {
+    public ReturnOrBorrowPresenter(ReturnOrBorrowViewModel returnOrBorrowViewModel, ViewManagerModel viewManagerModel, RentMenuViewModel rentMenuViewModel, ReturnBookViewModel returnBookViewModel, BorrowBookViewModel borrowBookViewModel, MainMenuViewModel mainMenuViewModel) {
         this.returnOrBorrowViewModel = returnOrBorrowViewModel;
         this.viewManagerModel = viewManagerModel;
         this.rentMenuViewModel = rentMenuViewModel;
         this.returnBookViewModel = returnBookViewModel;
         this.borrowBookViewModel = borrowBookViewModel;
+        this.mainMenuViewModel = mainMenuViewModel;
     }
 
     @Override
     public void returnBook() {
-
         System.out.println("Switched from Return or Borrow View to Return");
         viewManagerModel.setActiveView(returnBookViewModel.getViewName());
         viewManagerModel.firePropertyChanged();
@@ -38,7 +41,7 @@ public class ReturnOrBorrowPresenter implements RobOutputBoundary {
 
     @Override
     public void prepareCancelView() {
-        viewManagerModel.setActiveView(rentMenuViewModel.getViewName());
+        viewManagerModel.setActiveView(mainMenuViewModel.getViewName());
         viewManagerModel.firePropertyChanged();
         System.out.println("Switched from Return or Borrow View to rent Menu");
