+++ src/main/java/interface_adapter/returnorborrow/ReturnOrBorrowState.java
@@ -1,5 +1,14 @@
 package interface_adapter.returnorborrow;
 
 public class ReturnOrBorrowState {
+    private int bookId = 0;
 
+    public ReturnOrBorrowState(ReturnOrBorrowState copy) {
+        bookId = copy.bookId;
+    }
+    public ReturnOrBorrowState() {}
+
+    public  void setBookId(int bookId) {
+        this.bookId = bookId;
+    }
 }
